separated note card components
