+++ app/javascript/src/components/Dashboard/Notes/NotesCard/CardHeader.jsx
@@ -1,0 +1,22 @@
+import React from "react";
+
+import { MenuVertical } from "neetoicons";
+import { Dropdown } from "neetoui/v2";
+
+const CardHeader = ({ title, setShowModal }) => {
+  return (
+    <div className="flex justify-between w-full px-4 pt-4">
+      <h3>{title}</h3>
+      <Dropdown
+        icon={() => <MenuVertical size={20} color={"gray"} />}
+        position="bottom-start"
+        buttonStyle="text"
+      >
+        <li>Edit</li>
+        <li onClick={() => setShowModal(true)}>Delete</li>
+      </Dropdown>
+    </div>
+  );
+};
+
+export default CardHeader;
